requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
numpy==2.1.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.11.1
